+++ api/migrations/registry.py
@@ -12,6 +12,7 @@ from . import (
     repository_objects_v1,
     repository_objects_v2,
     repository_objects_v3,
+    repository_objects_v4,
     restore_history_v1,
     storage_objects_v1,
     storage_objects_v2,
@@ -23,6 +24,7 @@ MIGRATIONS = [
     repository_objects_v1,
     repository_objects_v2,
     repository_objects_v3,
+    repository_objects_v4,
     storage_objects_v1,
     storage_objects_v2,
 ]